+++ .gitignore
@@ -51,9 +51,14 @@ Module.symvers
 Mkfile.old
 dkms.conf
 
-# ft_ls
+# clion
 .idea/
-build/
 cmake-build-debug/
-debug
+build/
+
+# pvs
+.PVS-Studio/
+html
+
+# ft_ls
 ft_ls
