+++ src/args_parser.c
@@ -78,11 +78,11 @@ char		**files_parser(t_args *ls)
 
 static void	get_option(char alpha, t_opts *option)
 {
-	option->dot_files = (alpha == 'a') ? 1 : option->dot_files;
-	option->time_sort = (alpha == 't') ? 1 : option->time_sort;
-	option->long_format = (alpha == 'l') ? 1 : option->long_format;
-	option->reverse_sort = (alpha == 'r') ? 1 : option->reverse_sort;
-	option->recursive_read = (alpha == 'R') ? 1 : option->recursive_read;
+	option->dot_files = (alpha == OPTIONS[2]) ? 1 : option->dot_files;
+	option->time_sort = (alpha == OPTIONS[5]) ? 1 : option->time_sort;
+	option->long_format = (alpha == OPTIONS[3]) ? 1 : option->long_format;
+	option->reverse_sort = (alpha == OPTIONS[4]) ? 1 : option->reverse_sort;
+	option->recursive_read = (alpha == OPTIONS[1]) ? 1 : option->recursive_read;
 }
 
 /*
@@ -105,19 +105,18 @@ static int	is_option(char *str, t_opts *option)
 		while (str[i])
 		{
 			j = 1;
-			while (OPTIONS[j] != str[i] && j < ft_strlen(OPTIONS))
+			while (j < ft_strlen(OPTIONS) && str[i] != OPTIONS[j])
 				++j;
 			if (str[i] != OPTIONS[j])
 			{
-				ft_printf("ft_ls: illegal option -- %c\n", str[i]);
-				ft_printf("usage: ft_ls [%s] [file ...]\n", OPTIONS);
+				ft_printf(USE_MSG, str[i], OPTIONS);
 				exit(EXIT_FAILURE);
 			}
 			get_option(str[i++], option);
 		}
-		return (0);
+		return (EXIT_SUCCESS);
 	}
-	return (1);
+	return (EXIT_FAILURE);
 }
 
 /*
