+++ src/src.mk
@@ -7,6 +7,7 @@ SRC_SRCS	= main.c\
 			  sorting.c\
 			  output.c\
 			  long_format.c\
+			  long_format_2.c\
 			  free.c
 
 SRC			= $(addprefix $(SRC_DIR), $(SRC_SRCS))
